backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
httpx>=0.27.0
redis>=5.2.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.154.0
pytest>=8.3.0
